### 2019-12-01 #1924 @whedon
[PRE REVIEW]: OpenTripPlanner for R
**Submitting author:** @mem48 (<a href="http://orcid.org/0000-0002-9488-9183">Malcolm Morgan</a>)
**Repository:** <a href="https://github.com/ropensci/opentripplanner" target ="_blank">https://github.com/ropensci/opentripplanner</a>
**Version:** v0.2.0
