### 2019-12-02 #1924 @csoneson
Looks like this submission was [reviewed and accepted by rOpenSci](https://github.com/ropensci/software-review/issues/295).
