### 2019-12-02 #1924 @mem48
Hi @csoneson yes your right This package has been reviewed by rOpenSci: https://github.com/ropensci/software-review/issues/295, @ropensci/editors

@Kevin-Mattheus-Moerman I've just updated the .bib file with the missing DOI
