### 2019-12-02 #1926 @danielskatz
Note that this submission has been accepted by rOpenSci (https://github.com/ropensci/software-review/issues/295), so this JOSS process is just checking to ensure that JOSS criteria are met.
