### 2019-12-02 #1926 @danielskatz
👋 @mem48 - please merge the changes in https://github.com/ropensci/opentripplanner/pull/41 or let me know which you disagree with.
