### 2019-12-02 #1926 @danielskatz
Also in the second reference in the bib file, I see
```booktitle = {Algorithms ? {ESA} 2010},```

The ```?``` seems incorrect.  Can you check/fix this?
