### 2019-12-02 #1926 @danielskatz
The next steps are to make an archive of the code somewhere, potentially using Zenodo (as described [here](https://guides.github.com/activities/citable-code/)) or another archive, then let me know both the doi of the archive and the version number of the code, if it has changed.  Please also be sure the title and author metadata associated with the archive matches the paper title and authors.
