### 2019-12-02 #1926 @mem48
And both the paper and the archive are version 0.2
